# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BuildTrack Construction Expense Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/phases:
    get:
      operationId: listPhases
      tags: [phases]
      summary: List phases for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of phases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phase"
    post:
      operationId: createPhase
      tags: [phases]
      summary: Create a phase for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhaseBody"
      responses:
        "201":
          description: Created phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
  /phases/{id}:
    patch:
      operationId: updatePhase
      tags: [phases]
      summary: Update a phase
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhaseBody"
      responses:
        "200":
          description: Updated phase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phase"
    delete:
      operationId: deletePhase
      tags: [phases]
      summary: Delete a phase
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Phase deleted
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory items (optionally filter by project)
      parameters:
        - name: projectId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventory
      tags: [inventory]
      summary: Create an inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventory
      tags: [inventory]
      summary: Update an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    delete:
      operationId: deleteInventory
      tags: [inventory]
      summary: Delete an inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Inventory item deleted
...
components:
...
  schemas:
...
    Phase:
      type: object
      required: [id, projectId, name, status, createdAt, totalExpenses, expenseCount]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        totalExpenses: { type: number }
        expenseCount: { type: integer }
    CreatePhaseBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
    UpdatePhaseBody:
      type: object
      properties:
        name: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
...
    InventoryItem:
      type: object
      required: [id, name, unit, quantity, costPerUnit, reorderLevel, totalValue, isLow, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        unit: { type: string }
        quantity: { type: number }
        costPerUnit: { type: number }
        reorderLevel: { type: number }
        vendor: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        totalValue: { type: number }
        isLow: { type: boolean }
        createdAt: { type: string }
    CreateInventoryItemBody:
      type: object
      required: [name]
      properties:
        name: { type: string }
        unit: { type: string }
        quantity: { type: number }
        costPerUnit: { type: number }
        reorderLevel: { type: number }
        vendor: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
    UpdateInventoryItemBody:
      type: object
      properties:
        name: { type: string }
        unit: { type: string }
        quantity: { type: number }
        costPerUnit: { type: number }
        reorderLevel: { type: number }
        vendor: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
...